pyproject.toml
[tool.poetry]
name = "Aleksey94dan-gendiff"
version = "0.5.0"
description = ""
authors = ["Aleksey Danilchenko <danilchenko.aleksey94@gmail.com>"]
packages = [
    { include = "gendiff" }
]

[tool.poetry.dependencies]
python = "^3.7"

[tool.poetry.dev-dependencies]
pytest = "^5.3"
wemake-python-styleguide = "^0.13.2"
codeclimate-test-reporter = "^0.2.3"
pytest-cov = "=2.4.0"
coverage = "=4.3.4"
PyYAML = "^5.2"
pylint = "^2.4"

[[tool.poetry.source]]
name = "test"
url = "https://test.pypi.org/legacy/"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

